Create basic GUI layout and bind in a basic controller
This is as far as I got with the GUI so far. The idea is to be able
to either start a new recording or load a past one and then process
the data. The incoming signal, fft, and correlation could be shown
in the graphs. A placeholder circle was added to display detection
direction relative to the front of the array.